@@ -5,41 +5,31 @@
   "author": "ivmartel",
   "license": "GPL-3.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "start": "vue-cli-service serve",
-    "lint": "vue-cli-service lint",
-    "test": "vue-cli-service test:unit --watch",
-    "test-ci": "vue-cli-service test:unit",
-    "e2e": "vue-cli-service test:e2e",
-    "build": "vue-cli-service build",
-    "build-trunk": "NODE_ENV=production DEPLOY_NAME=trunk vue-cli-service build",
-    "build-stable": "NODE_ENV=production DEPLOY_NAME=stable vue-cli-service build"
+    "dev": "vite",
+    "build": "vite build",
+    "build-trunk": "vite build --mode development --base /dwv-vue/demo/trunk",
+    "build-stable": "vite build --mode production --base /dwv-vue/demo/stable",
+    "preview": "vite preview",
+    "test": "vitest",
+    "test-ci": "vitest run",
+    "lint": "eslint src/**/*.{js,vue}"
   },
   "dependencies": {
-    "core-js": "^3.37.1",
     "dwv": "0.33.0",
-    "register-service-worker": "^1.7.2",
-    "vue": "~2.7.16",
-    "vue-material": "~1.0.0-beta"
+    "vue": "^3.4.27"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.7",
-    "@babel/eslint-parser": "^7.24.1",
-    "@vue/cli-plugin-babel": "~5.0.8",
-    "@vue/cli-plugin-e2e-nightwatch": "~5.0.8",
-    "@vue/cli-plugin-eslint": "~5.0.8",
-    "@vue/cli-plugin-pwa": "~5.0.8",
-    "@vue/cli-plugin-unit-jest": "~5.0.8",
-    "@vue/cli-service": "~5.0.8",
-    "@vue/test-utils": "^1.3.6",
-    "@vue/vue2-jest": "^27.0.0",
-    "babel-jest": "^27.5.1",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vue/test-utils": "^2.4.6",
     "canvas": "^2.11.2",
-    "chromedriver": "^125.0.3",
-    "eslint": "^8.57.0",
-    "eslint-plugin-vue": "^9.18.1",
-    "geckodriver": "^4.4.1",
-    "jest": "^27.1.0",
-    "vue-template-compiler": "^2.7.16"
+    "eslint": "^9.4.0",
+    "eslint-plugin-vue": "^9.26.0",
+    "jsdom": "^24.1.0",
+    "vite": "^5.2.0",
+    "vite-plugin-pwa": "^0.20.0",
+    "vitest": "^1.6.0",
+    "vuetify": "^3.6.9"
   }
 }
