@@ -1,11 +1,25 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
+import './style.css'
+
+// Vuetify
+import 'vuetify/styles'
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+import { aliases, md } from 'vuetify/iconsets/md'
+
 import App from './App.vue'
-import './registerServiceWorker'
 
-import 'vue-material/dist/vue-material.min.css'
+const vuetify = createVuetify({
+  components,
+  directives,
+  icons: {
+    defaultSet: 'md',
+    aliases,
+    sets: {
+      md,
+    },
+  },
+})
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).use(vuetify).mount('#app')
